requirements.txt
# ==============================================================================
# V-CoT Diffusion Policy - Requirements
# ==============================================================================
# Note: Do not blindly upgrade packages in the "Simulation & Robotics" section, 
# as they are heavily dependent on legacy API structures (e.g., Gym < 0.26).

# --- Core PyTorch & CUDA 12.1 ---
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# --- V-CoT Module (High-level Reasoning) ---
# Pinning for Python 3.8 compatibility to avoid PEFT/Diffusers version hell
diffusers<0.31.0
transformers<4.45.0
accelerate>=0.21.0
peft==0.13.2
bitsandbytes
sentencepiece
protobuf
# [ADDED] Required for VLM dataloading and efficient weight serialization
datasets>=2.14.0
safetensors

# --- Diffusion Policy Core (Low-level Control) ---
dill
hydra-core>=1.2
hydra-colorlog
omegaconf
zarr
numcodecs
wandb
einops
tqdm
termcolor
scikit-image
scikit-learn
shapely
networkx

# --- Simulation & Robotics (CAUTION: Strictly Pinned Versions) ---
mujoco>=3.1.0
dm_control>=1.0.14
robosuite==1.4.1
robomimic

# [CRITICAL FIX]
# gym==0.24.1: Diffusion Policy expects `obs = env.reset()`. 
# Gym >=0.26 returns a tuple `(obs, info)`, causing concatenation failures.
gym==0.24.1

# [CRITICAL FIX]
# numpy==1.23.5: Gym 0.24.1 relies on legacy random generator methods.
# NumPy >= 1.24 removes these methods, throwing `_generator_ctor` errors.
numpy==1.23.5
gymnasium

# --- Data Visualization & Utilities ---
matplotlib
seaborn
imageio[ffmpeg]
opencv-python
h5py
numba
